consumer_notification_recorder/pyproject.toml
[tool.poetry]
name = "consumer-messages"
version = "0.1.0"
description = ""
authors = ["NikolaySolop <nikolay.solop@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "^2.6.1"
pika = "^1.3.2"
aio-pika = "^9.4.3"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
